datasets/generation/requirements.txt
openai
psycopg2-binary
python-dotenv
tenacity